+++ src/client/redux/greeting/actions.js
@@ -1,3 +1,4 @@
+import axios from 'axios'
 import * as type from './types'
 
 export const requestGreeting = () => {
@@ -6,25 +7,27 @@ export const requestGreeting = () => {
   }
 }
 
-const getGreetingError = () => Observable.of({
-  type: type.GET_GREETING_ERROR,
-  error: 'Loading Error.'
-})
+const getGreetingError = (error) => {
+  return {
+    type: type.GET_GREETING_ERROR,
+    error: error
+  }
+}
 
 const getGreetingSuccess = (payload) => {
   let hr = new Date().getHours()
   let greeting
 
   if (hr >= 0 && hr <= 12) {
-    greeting = payload.greetings[0]
+    greeting = payload.data.greetings[0]
   }
 
   if (hr >= 12 && hr <= 18) {
-    greeting = payload.greetings[1]
+    greeting = payload.data.greetings[1]
   }
 
   if (hr >= 18 && hr <= 24) {
-    greeting = payload.greetings[2]
+    greeting = payload.data.greetings[2]
   }
 
   return {
@@ -33,10 +36,13 @@ const getGreetingSuccess = (payload) => {
   }
 }
 
-export const greetingEpic = action$ =>
-  action$.ofType(type.GET_GREETING)
-    .mergeMap(action =>
-      ajax.getJSON('./data/greeting.json')
-        .map(response => getGreetingSuccess(response))
-        .catch(getGreetingError)
-    )
+export const greetingEpic = action$ => {
+  return (
+    action$.ofType(type.GET_GREETING)
+      .mergeMap(action => {
+        return axios.get('./data/greeting.json')
+          .then(res => getGreetingSuccess(res))
+          .catch(error => getGreetingError(error))
+      })
+  )
+}
